Split the sync module into server/sync/
One 355-line file stacked four concerns that change for different
reasons and shared nothing but the lease: holding the lease, choosing
which connections are owed, polling one connection, and running the hop
that ties them together. They are now three modules in a straight line —
run-sync-hop depends on poll-connection depends on sync-lease, no
back-edges — behind a barrel that re-exports exactly what server.ts
exported before.

The tests split with them: the lease suite and the query-shape, polling
window and backoff-ladder cases move to the module they exercise
directly. The behaviours that are only observable by running a hop stay
together in run-sync-hop.test.ts rather than being rewritten to call
pollConnection — every assertion moves verbatim, and the count is
unchanged at 55.
