+++ apps/app/src/features/integrations/server/sync/index.ts
@@ -1,0 +1,6 @@
+export { runSyncHop } from "./run-sync-hop";
+export {
+  acquireSyncLease,
+  continueSyncLease,
+  type SyncLease,
+} from "./sync-lease";
